move ColumnRegExpMetric data classes to the metric module
