+++ src/evidently/metrics/data_integrity_metrics.py
@@ -165,35 +165,6 @@ class DataIntegrityMetricsRenderer(MetricRenderer):
 
 
 @dataclass
-class DataIntegrityValueByRegexpStat:
-    """Statistics about matched by a regular expression values in a column for one dataset"""
-
-    # count of matched values in the column, without NaNs
-    number_of_matched: int
-    # count of not matched values in the column, without NaNs
-    number_of_not_matched: int
-    # count of rows in the column, including matched, not matched and NaNs
-    number_of_rows: int
-    # map with matched values (keys) and count of the values (value)
-    table_of_matched: Dict[str, int]
-    # map with not matched values (keys) and count of the values (values)
-    table_of_not_matched: Dict[str, int]
-
-
-@dataclass
-class DataIntegrityValueByRegexpMetricResult:
-    # name of the column that we check by the regular expression
-    column_name: str
-    # the regular expression as a string
-    reg_exp: str
-    top: int
-    # match statistic for current dataset
-    current: DataIntegrityValueByRegexpStat
-    # match statistic for reference dataset, equals None if the reference is not present
-    reference: Optional[DataIntegrityValueByRegexpStat] = None
-
-
-@dataclass
 class DataIntegrityNullValuesStat:
     """Statistics about null values in a dataset"""
